feat: Event Bridge 통합 및 알람 시스템 구현
- NATS 이벤트 수신 → DB 저장 기능 추가 (event.updated.{stream_id})
- 알람 서비스 통합 (alarm.updated 토픽 구독)
- 4가지 알람 프로토콜 구현:
  - Adam6050 (Modbus TCP I/O)
  - IoLogik_E1211 (MOXA REST API)
  - AEPEL_IPSpeaker (HTTP 스피커)
  - LA6_POE (PATLITE LED 타워)
- 센서/센서타입 CRUD API 추가 (/api/v2/sensors)
- NATS wakeup 서비스 추가 (inference.start 응답)
- 시작 시 기본 센서 타입 자동 시드

+++ app/models/event.py
@@ -3,7 +3,7 @@
 import json
 from typing import Any
 
-from sqlalchemy import BigInteger, Index, String, Text
+from sqlalchemy import BigInteger, Index, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.base import Base
@@ -14,7 +14,8 @@ class Event(Base):
 
     __tablename__ = "events"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    # SQLite requires INTEGER (not BIGINT) for autoincrement
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
 
     # Event identification
     event_setting_id: Mapped[str | None] = mapped_column(String(255), nullable=True)
